+++ stripe/src/main/java/com/stripe/android/checkout/CheckoutViewModel.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.view
+package com.stripe.android.checkout
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
